+++ common/src/main/java/me/pandamods/fallingtrees/config/client/AnimationConfig.java
@@ -1,0 +1,7 @@
+package me.pandamods.fallingtrees.config.client;
+
+public class AnimationConfig {
+	public float fallAnimLength = 2.5f;
+	public float bounceAngleHeight = 10;
+	public float bounceAnimLength = 1;
+}