+++ common/src/main/java/me/pandamods/fallingtrees/config/client/SoundSettingsConfig.java
@@ -1,0 +1,7 @@
+package me.pandamods.fallingtrees.config.client;
+
+public class SoundSettingsConfig {
+	public boolean enabled = true;
+	public float startVolume = 0.7f;
+	public float endVolume = 0.7f;
+}