+++ common/src/main/java/me/pandamods/fallingtrees/config/common/FeaturesConfig.java
@@ -1,0 +1,7 @@
+package me.pandamods.fallingtrees.config.common;
+
+public class FeaturesConfig {
+	public boolean disableCactusTrees = false;
+	public boolean disableBambooTrees = false;
+	public boolean disableChorusTrees = false;
+}