+++ common/src/main/java/me/pandamods/fallingtrees/mixin/PlayerMixin.java
@@ -4,6 +4,7 @@ import me.pandamods.fallingtrees.FallingTrees;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Pose;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;