+++ Changelog.md
@@ -1,3 +1,13 @@
 ## Version 0.9 Changelog
 * **What's been fixed**
   * Server crash because client side only code is tried to get executed on the server
+  * Common config not always syncing with singleplayer 
+
+
+* **What's been recoded**
+  * Animation
+
+
+* **What's been added**
+  * Animation config options
+  * Tree life length config option